backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
requests
google-genai
playwright
